Initial normalization Implementation (#3)
* Initial implementation

* Add link to README.md

* Features: normlaize Routines calls

* Build(.gitignore): exclude the directory of Exercism data

* Feature(representer): ignore style of identifiers

This change included:
 - implemented alias type for string with `hash` and `==`
 - simplifying table with alias type

* Feature(execution): define path with -d:

The path to the file with -d:path=<FILE>

* Fix(mapping): suppress unused warning

* Feature(representer): normalize import, exclude and discard statements

* Docs(representer): Fix comments typos and style

* Fix(config): output executable to bin/ dir

Output the executable to the bin/ directory as per spec and to avoid
`represeneter.out` file

* Fix:(call normalization): fix normalization for DotExpr

This changes the normalization to the correct proc call for different
types of call

* Fix(value normalization): Fix normalization of value

The normalization of value should never define a new variable to be
"placeheld". Therefore this commit ensures that if anything is
"placeheld" it has to have been previously added to the `map`

* Fix(mapping): Fix equality of Identfiers

Change `==` for `NormalizedIdent` to use `string`'s `cmpIgnoreStyle`
proc to avoid recursion due to ambiguous `==` call as suggested by @ee7
in PR #3 

Co-authored-by: ee7 <45465154+ee7@users.noreply.github.com>

* Fix(representer): normalize each side of `DotExpr` separately

There will probably be some issues with this
For example if there is a variable `echo`, it's ambiguous whether to
normalize or not

* Feature(representer): add normalization for expresssion with parentheses

* Fix(mapping): change approach to ident comparison

Incorporated suggestions from @ee7 to use `distint string` and
borrow/implement the neccesary procs. The previous approach did not work
so switching seemed like the only option.
The downside is that `string`s have to be converted to
`NormalizedIdent`s.

Co-authored-by: ee7 <45465154+ee7@users.noreply.github.com>

* Test(representer): initial tests for representer

* Test(mapping): tests for ident mapping table

* Fix(representer): don't normalized named parameter

Instead of removing the variable name for the named parameters, the
variable name is now normalized if it has previously been seen

* Test(representer): add test for name parameter

* Style: fix test names, remove trailing whitespace & add final newline

Co-authored-by: ee7 <45465154+ee7@users.noreply.github.com>

* Fix(mapping): implement json conversion for `IdentMap`

* Refactor(mapping): inline conversion of `NormalizedIden[]` to `string[]`

* Refactor(representer): use `result` for return value

As suggested by @ee7, the `result` variable is more explicit than just implicit return or different variable

Co-authored-by: ee7 <45465154+ee7@users.noreply.github.com>