Feature: use nimscripter
This commit upgrades the representer to use
https://github.com/beef331/nimscripter.

This module is a branch between Nimscript (a subset of nim that runs in
the VM) and the compiled nim. It embeds an interpreter in to the
executable, and doesn't require a nim compiler to be in path. The
primary advantage is the the bridge is built-in, so there is no need for
recompiling to create a representation for  every exercise, as the
representer uses nim's macros which only function in the VM. Without
this, only the VM is used, but the executable has to be recompiled every
time.

This also greatly simplifies the testing process, as the representation
creation doesn't need to happen at compile to and injecting in. Rather,
it can happen at runtime with the VM bridge.