Feature: upgrade macro evaluation strategy to use `nimscripter` (#37)
This commit upgrades the representer to use https://github.com/beef331/nimscripter.

This module is a branch between Nimscript (a subset of nim that runs in the VM) and the compiled nim. It embeds an interpreter in to the executable, and doesn't require a nim compiler to be in path. The primary advantage is the the bridge is built-in, so there is no need for recompiling to create a representation for  every exercise, as the representer uses nim's macros which only function in the VM. Without this, only the VM is used, but the executable has to be recompiled every time.

This also greatly simplifies the testing process, as the representation creation doesn't need to happen at compile to and injecting in. Rather, it can happen at runtime with the VM bridge.

This also add the use of https://github.com/docopt/docopt.nim form command line parsing. This was not possible before, because everything happened at compile time and there was no code being run.

CI:

https://github.com/jiro4989/setup-nim-action is used to install nim using choosenim with the desired version.

Caching is utilized to not have to redownload the dependent packages again and invalidates the cache when the nimble file changes. This is the example used on the `setup-nim-action` repo `README.md`

Closes #9 as not relevant

Co-authored-by: Erik Schierboom <erik_schierboom@hotmail.com>
Co-authored-by: ee7 <45465154+ee7@users.noreply.github.com>